# server-application/docs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Reds App API: Swagger Documentation - OpenAPI 3.0",
    "description": "This is a documentation page for the Reds App server-side API created by IFB398/IFB399 Team 125 based on the OpenAPI 3.0 specification.\n\nLinks:\n- [GitHub Repository](https://github.com/VmiteOCE/IFB398-T125)",
    "termsOfService": "https://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.27"
  },
...
  "servers": [
    {
      "url": "https://localhost:3000/",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/user": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Create a new user.",
        "description": "This can only be done by a logged in admin user.",
        "operationId": "createUser",
        "requestBody": {
          "$ref": "#/components/requestBodies/UserCreateRequestBody"
        },
        "responses": {
          "201": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreateResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database error"
          }
        }
      },
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get a list of all users.",
        "description": "This can only be done by a logged in admin user.",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUsersResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database error"
          }
        }
      },
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "Update password or role of an existing user.",
        "description": "This can only be done by a logged in admin user. Intended for use on a user management page.",
        "operationId": "updateUser",
        "requestBody": {
          "$ref": "#/components/requestBodies/UserUpdateRequestBody"
        },
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserUpdateResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database error"
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Delete an existing user.",
        "description": "This can only be done by a logged in admin user.",
        "operationId": "deleteUser",
        "requestBody": {
          "$ref": "#/components/requestBodies/UserDeleteRequestBody"
        },
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDeleteResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database error"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Logs user into the system.",
        "description": "Log an existing user into the system.",
        "operationId": "loginUser",
        "requestBody": {
          "$ref": "#/components/requestBodies/UserLoginRequestBody"
        },
        "responses": {
          "201": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserLoginResponse"
                }
              }
            }
          },
          "500": {
            "description": "Login error"
          }
        },
        "security": []
      }
    },
    "/user/password": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update password for the logged in user.",
        "description": "Only available to logged in users.",
        "operationId": "updatePassword",
        "requestBody": {
          "$ref": "#/components/requestBodies/UpdatePasswordRequestBody"
        },
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePasswordResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database error"
          }
        }
      }
    },
    "/user/keybinds": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Retrieves keybindings for the logged in user.",
        "description": "Returns a JSON object containing a full keymap if the user has modified any keybindings. If no keybinds have been altered, keybinds returns as null.",
        "operationId": "getKeybinds",
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetKeybindsResponse"
                },
                "examples": {
                  "keybindsFound": {
                    "summary": "Custom Keybinds Saved",
                    "$ref": "#/components/schemas/Keybinds"
                  },
                  "noKeybinds": {
                    "value": {
                      "error": false,
                      "keybinds": null
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error"
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Updates keybindings for the logged in user.",
        "description": "Takes a full keymap and saves the cusom keybinds for the logged in user.",
        "operationId": "updateKeybinds",
        "requestBody": {
          "$ref": "#/components/requestBodies/UpdateKeybindsRequestBody"
        },
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateKeybindsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database error"
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Resets keybindings for the logged in user to default.",
        "description": "Resets any existing keybinds, reverting the user to default settings.",
        "operationId": "resetKeybinds",
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetKeybindsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "UserRole": {
        "type": "object",
        "required": [
          "username",
          "role"
        ],
        "properties": {
          "username": {
            "type": "string",
            "description": "Username of the account",
            "example": "admin"
          },
          "role": {
            "type": "string",
            "description": "User access role",
            "enum": [
              "admin",
              "editor",
              "viewer"
            ],
            "example": "admin"
          }
        }
      },
      "UserLogin": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "description": "Username to authenticate",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "description": "Password to authenticate",
            "example": "admin"
          }
        }
      },
      "UserFull": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UserLogin"
          },
          {
            "$ref": "#/components/schemas/UserRole"
          }
        ],
        "example": {
          "username": "test",
          "password": "test",
          "role": "viewer"
        }
      },
      "JWTPayload": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UserRole"
          },
          {
            "type": "object",
            "required": [
              "iat",
              "exp"
            ],
            "properties": {
              "iat": {
                "type": "integer",
                "description": "Issued at timestamp",
                "example": 1787390825
              },
              "exp": {
                "type": "integer",
                "description": "Expiration timestamp",
                "example": 1787477225
              }
            }
          }
        ]
      },
      "UserLoginResponse": {
        "type": "object",
        "required": [
          "error",
          "message",
          "token_type",
          "token",
          "token_payload"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Login successful"
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          },
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImFkbWluIiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNzg3MzkwODI1LCJleHAiOjE3ODc0NzcyMjV9.BKKd_DX2a2DknyHBQq8lZxCcHwD5o3TfDLst0zHY6hg"
          },
          "token_payload": {
            "$ref": "#/components/schemas/JWTPayload"
          }
        }
      },
      "UserCreateResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "User 'test' created successfully with role 'viewer'."
          }
        }
      },
      "GetUsersResponse": {
        "type": "object",
        "required": [
          "error",
          "users"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "example": false
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserRole"
            }
          }
        },
        "example": {
          "error": false,
          "users": [
            {
              "username": "admin",
              "role": "admin"
            },
            {
              "username": "test",
              "role": "viewer"
            }
          ]
        }
      },
      "UserUpdate": {
        "type": "object",
        
        "required": [
          "username"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "test"
          }
        },
        "oneOf": [
          {
            "title": "Update Password Only",
            "type": "object",
            "required": [
              "password"
            ],
            "properties": {
              "password": {
                "type": "string",
                "example": "newpassword"
              }
            }
          },
          {
            "title": "Update Role Only",
            "type": "object",
            "required": [
              "role"
            ],
            "properties": {
              "role": {
                "type": "string",
                "enum": [
                  "admin",
                  "editor",
                  "viewer"
                ],
                "example": "editor"
              }
            }
          },
          {
            "title": "Update Password & Role",
            "type": "object",
            "required": [
              "password",
              "role"
            ],
            "properties": {
              "password": {
                "type": "string",
                "example": "newpassword"
              },
              "role": {
                "type": "string",
                "enum": [
                  "admin",
                  "editor",
                  "viewer"
                ],
                "example": "editor"
              }
            }
          }
        ]
      },
      "UserUpdateResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "User 'test' updated successfully."
          }
        }
      },
      "UserDelete": {
        "type": "object",
        "required": [
          "username"
        ],
        "properties": {
          "username": {
            "$ref": "#/components/schemas/UserRole/properties/username"
          }
        },
        "example": {
          "username": "test"
        }
      },
      "UserDeleteResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "User 'test' deleted successfully."
          }
        }
      },
...
      "UpdatePasswordResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Password updated successfully."
          }
        }
      },
      "Keybinds": {
        "type": "object",
        "description": "Mapping of game events and controls to QWERTY keyboard keys",
        "required": [
          "Pass",
          "Kick",
          "Catch",
          "Ruck",
          "Scrum",
          "Penalty",
          "Advantage",
          "Turnover",
          "Lineout",
          "Maul",
          "Pause",
          "Add_Time",
          "Remove_Time",
          "Move_Zone_Left",
          "Move_Zone_Right",
          "Swap_Team",
          "Swap_Direction"
        ],
        "properties": {
          "Pass": {
            "type": "string",
            "example": "P"
          },
          "Kick": {
            "type": "string",
            "example": "K"
          },
          "Catch": {
            "type": "string",
            "example": "C"
          },
          "Ruck": {
            "type": "string",
            "example": "R"
          },
          "Scrum": {
            "type": "string",
            "example": "S"
          },
          "Penalty": {
            "type": "string",
            "example": "E"
          },
          "Advantage": {
            "type": "string",
            "example": "A"
          },
          "Turnover": {
            "type": "string",
            "example": "T"
          },
          "Lineout": {
            "type": "string",
            "example": "L"
          },
          "Maul": {
            "type": "string",
            "example": "M"
          },
          "Pause": {
            "type": "string",
            "example": "Space"
          },
          "Add_Time": {
            "type": "string",
            "example": "="
          },
          "Remove_Time": {
            "type": "string",
            "example": "-"
          },
          "Move_Zone_Left": {
            "type": "string",
            "example": "Left Arrow"
          },
          "Move_Zone_Right": {
            "type": "string",
            "example": "Right Arrow"
          },
          "Swap_Team": {
            "type": "string",
            "example": "Tab"
          },
          "Swap_Direction": {
            "type": "string",
            "example": "Up Arrow"
          }
        },
        "example": {
          "Pass": "P",
          "Kick": "K",
          "Catch": "C",
          "Ruck": "R",
          "Scrum": "S",
          "Penalty": "E",
          "Advantage": "A",
          "Turnover": "T",
          "Lineout": "L",
          "Maul": "M",
          "Pause": "Space",
          "Add_Time": "=",
          "Remove_Time": "-",
          "Move_Zone_Left": "Left Arrow",
          "Move_Zone_Right": "Right Arrow",
          "Swap_Team": "Tab",
          "Swap_Direction": "Up Arrow"
        }
      },
      "GetKeybindsResponse": {
        "type": "object",
        "required": [
          "error",
          "keybinds"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "example": false
          },
          "keybinds": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Keybinds"
              }
            ]
          }
        }
      },
      "UpdateKeybindsResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Keybinds updated successfully."
          }
        }
      },
      "ResetKeybindsResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Keybinds successfully reset to default."
          }
        }
      }
    },
    "requestBodies": {
...
      "UserLoginRequestBody": {
        "description": "Payload required to authenticate a user",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UserLogin"
            }
          },
          "application/x-www-form-urlencoded": {
            "schema": {
              "$ref": "#/components/schemas/UserLogin"
            }
          }
        }
      },
      "UserCreateRequestBody": {
        "description": "Payload required to create a new user",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UserFull"
            }
          },
          "application/x-www-form-urlencoded": {
            "schema": {
              "$ref": "#/components/schemas/UserFull"
            }
          }
        }
      },
      "UserUpdateRequestBody": {
        "description": "Payload required to update a user's password or role. Must supply username and at least password or role.",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UserUpdate"
            },
            "examples": {
              "updateRoleOnly": {
                "summary": "Option 1: Update Role Only",
                "description": "Updates only the user's role",
                "value": {
                  "username": "test",
                  "role": "editor"
                }
              },
              "updatePasswordOnly": {
                "summary": "Option 2: Update Password Only",
                "description": "Updates only the user's password",
                "value": {
                  "username": "test",
                  "password": "newpassword"
                }
              },
              "updateBoth": {
                "summary": "Option 3: Update Password and Role",
                "description": "Updates both the password and role simultaneously",
                "value": {
                  "username": "test",
                  "password": "newpassword",
                  "role": "editor"
                }
              }
            }
          },
          "application/x-www-form-urlencoded": {
            "schema": {
              "$ref": "#/components/schemas/UserUpdate"
            },
            "examples": {
              "updatePasswordOnly": {
                "summary": "Option 1: Update Password Only",
                "value": {
                  "username": "test",
                  "password": "newSecurePassword123"
                }
              },
              "updateRoleOnly": {
                "summary": "Option 2: Update Role Only",
                "value": {
                  "username": "test",
                  "role": "editor"
                }
              },
              "updateBoth": {
                "summary": "Option 3: Update Password and Role",
                "value": {
                  "username": "test",
                  "password": "newSecurePassword123",
                  "role": "editor"
                }
              }
            }
          }
        }
      },
      "UserDeleteRequestBody": {
        "description": "Payload required to delete a user account",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UserDelete"
            }
          },
          "application/x-www-form-urlencoded": {
            "schema": {
              "$ref": "#/components/schemas/UserDelete"
            }
          }
        }
      },
      "UpdatePasswordRequestBody": {
        "description": "Payload required to update the password of a logged in user",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UpdatePassword"
            }
          },
          "application/x-www-form-urlencoded": {
            "schema": {
              "$ref": "#/components/schemas/UpdatePassword"
            }
          }
        }
      },
      "UpdateKeybindsRequestBody": {
        "description": "Payload required to update user keybindings",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Keybinds"
            }
          },
          "application/x-www-form-urlencoded": {
            "schema": {
              "$ref": "#/components/schemas/Keybinds"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token directly (e.g. eyJhbGciOiJKV1Qi...)"
      }
    }
  }
}